refactor: make separate component for upload form, change error dialog component
