+++ src/entities/photo/index.ts
@@ -1,2 +1,3 @@
 export * from './model/service/photoApiSlice'
+export { AddProfilePhotoDialog } from './ui/AddProfilePhotoDialog'
 export { Photo } from './ui/Photo'
