+++ src/entities/photo/ui/Photo.module.scss
@@ -61,3 +61,10 @@
     background-color: var(--danger-color-700);
   }
 }
+
+/* Photo dialog */
+.dialog {
+  overflow-y: auto;
+}
+
+/* ============ */
