+++ src/features/profile/uploadUserPhoto/index.ts
@@ -1,0 +1,1 @@
+export { UploadPhoto } from './ui/UploadPhoto'
