+++ src/features/profile/uploadUserPhoto/ui/UploadPhoto.module.scss
@@ -1,7 +1,3 @@
-.dialog {
-  overflow-y: auto;
-}
-
 .photo {
   padding: 0 3.5rem;
 }
